refactoring dummy router
