+++ src/routes/home.js
@@ -1,30 +1,9 @@
 const express = require("express");
-const path = require("path");
-const fs = require("fs");
-const moment = require("moment");
+
+const { getIndex } = require("../controllers/home");
 
 const router = express.Router();
 
-const homePath = path.join(__dirname, "../../data/home.json");
+router.get("/", getIndex);
 
-function prepareHome(data) {
-  const result = JSON.parse(data);
-
-  result.calendar.forEach(element => {
-    const weekday = moment(element.date, "DD.MM.YYYY")
-      .format("dddd")
-      .toLowerCase();
-    element.dayicon = `img/calendar/${weekday}.png`;
-  });
-
-  return result;
-}
-
-router.get("/", (req, res) => {
-  fs.readFile(homePath, (err, data) => {
-    if (err) throw err;
-    res.render("home", prepareHome(data));
-  });
-});
-
-module.exports = { prepareHome, router };
+module.exports = { router };
